[hotfix] Make SecurityUtils.SecurityConfiguration a toplevel class
